Rename files in accordance with airbnb style guide naming conventions
